Moved styled components to own files
