+++ client/src/dropdown-items/ResultsDropdown.js
@@ -1,20 +1,18 @@
 import { arrayOf, object, string } from 'prop-types';
 import React from 'react';
-import Dropdown from 'react-bootstrap/Dropdown';
-import styled from 'styled-components';
-import dropdownItems from './dropdown-items';
-import StyledDropdownItem from './dropdown-items/StyledDropdownItem';
+import dropdownItems from '.';
+import { StyledDropdown, StyledDropdownItem } from './styled';
 
 const ResultsDropdown = ({ results, activeKey }) => {
   const { textExtractor } = dropdownItems[activeKey];
   return (
-    <Dropdown>
+    <StyledDropdown>
       {results.map(result => (
         <StyledDropdownItem key={result._id} eventKey={result._id}>
           {textExtractor(result)}
         </StyledDropdownItem>
       ))}
-    </Dropdown>
+    </StyledDropdown>
   );
 };
 
@@ -27,7 +25,4 @@ ResultsDropdown.defaultProps = {
   results: [],
 };
 
-export default styled(ResultsDropdown)`
-  position: fixed;
-  width: 400px;
-`;
+export default ResultsDropdown;
