+++ client/src/dropdown-items/styled.js
@@ -1,7 +1,12 @@
 import Dropdown from 'react-bootstrap/Dropdown';
 import styled from 'styled-components';
 
-export default styled(Dropdown.Item)`
+export const StyledDropdown = styled(Dropdown)`
+  position: fixed;
+  width: 400px;
+`;
+
+export const StyledDropdownItem = styled(Dropdown.Item)`
   height: 50px;
   display: flex;
   background-color: white;
